+++ CHANGELOG.md
@@ -8,12 +8,15 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 ## v2.12.0-dev.3
 ### Added
 - Added Settings option "Console: History Size" to adjust the number of lines stored in the console history when running packages. Defaults to 9001 lines.
+#### Model Browser
+- Added AuraFlow & Flux base model types to the CivitAI model browser
 #### Checkpoint Manager
 - Added "New Directory" and "Delete" options to the context menu of the tree view.
 - Added new toggle for drag & drop - when enabled, all selected models will now move together with the dragged model
 - Added "File Size" sorting option
 - Added "Hide Empty Categories" toggle
 - Added "Select All" button to the InfoBar (shown when at least one model is selected)
+- Added "unet" shared model folder for ComfyUI
 ### Changed
 - The "Download Failed" message for model downloads is now persistent until dismissed
 ### Fixed